+++ src/containerapp/azext_containerapp/tests/latest/test_containerapp_java_component.py
@@ -5,7 +5,6 @@
 
 from azure.cli.command_modules.containerapp._utils import format_location
 
-from azure.cli.testsdk.scenario_tests import AllowLargeResponse
 from azure.cli.testsdk import (ScenarioTest, ResourceGroupPreparer, JMESPathCheck)
 
 from .common import TEST_LOCATION, STAGE_LOCATION
@@ -13,7 +12,6 @@ from .utils import create_containerapp_env
 
 
 class ContainerappJavaComponentTests(ScenarioTest):
-    @AllowLargeResponse(8192)
     @ResourceGroupPreparer()
     def test_containerapp_java_component_deprecated(self, resource_group):
         # type "linkers" is not available in North Central US (Stage), if the TEST_LOCATION is "northcentralusstage", use francecentral as location
@@ -28,6 +26,8 @@ class ContainerappJavaComponentTests(ScenarioTest):
         eureka_name = "myeureka"
 
         create_containerapp_env(self, env_name, resource_group)
+        env = self.cmd('containerapp env show -g {} -n {}'.format(resource_group, env_name)).get_output_in_json()
+        default_domain=env['properties']['defaultDomain']
 
         # List Java Components
         java_component_list = self.cmd("containerapp env java-component list -g {} --environment {}".format(resource_group, env_name)).get_output_in_json()
@@ -52,7 +52,7 @@ class ContainerappJavaComponentTests(ScenarioTest):
         java_component_list = self.cmd("containerapp env java-component list -g {} --environment {}".format(resource_group, env_name)).get_output_in_json()
         self.assertTrue(len(java_component_list) == 2)
 
-        # Update Config & Eureka
+        # Update Java Components
         self.cmd(
             'containerapp env java-component spring-cloud-config update -g {} -n {} --environment {} --configuration spring.cloud.config.server.git.uri=https://github.com/Azure-Samples/piggymetrics-config.git'.format(
                 resource_group, config_name, env_name), checks=[
@@ -64,10 +64,11 @@ class ContainerappJavaComponentTests(ScenarioTest):
         self.cmd('containerapp env java-component spring-cloud-eureka update -g {} -n {} --environment {} --configuration'.format(resource_group, eureka_name, env_name), checks=[
                 JMESPathCheck('name', eureka_name),
                 JMESPathCheck('properties.componentType', "SpringCloudEureka"),
+                JMESPathCheck('properties.ingress.fqdn', eureka_name + "-azure-java.ext." + default_domain),
                 JMESPathCheck('properties.provisioningState', "Succeeded")
         ])
 
-        # Show Config & Eureka
+        # Show Java Components
         self.cmd('containerapp env java-component spring-cloud-config show -g {} -n {} --environment {}'.format(resource_group, config_name, env_name), checks=[
             JMESPathCheck('name', config_name),
             JMESPathCheck('properties.componentType', "SpringCloudConfig"),
@@ -77,6 +78,7 @@ class ContainerappJavaComponentTests(ScenarioTest):
         self.cmd('containerapp env java-component spring-cloud-eureka show -g {} -n {} --environment {}'.format(resource_group, eureka_name, env_name), checks=[
             JMESPathCheck('name', eureka_name),
             JMESPathCheck('properties.componentType', "SpringCloudEureka"),
+            JMESPathCheck('properties.ingress.fqdn', eureka_name + "-azure-java.ext." + default_domain),
             JMESPathCheck('properties.provisioningState', "Succeeded")
         ])
 
@@ -104,7 +106,7 @@ class ContainerappJavaComponentTests(ScenarioTest):
                 JMESPathCheck('length(properties.template.serviceBinds)', 2)
         ])
 
-        # Delete Config & Eureka
+        # Delete Java Components
         self.cmd('containerapp env java-component spring-cloud-config delete -g {} -n {} --environment {} --yes'.format(resource_group, config_name, env_name), expect_failure=False)
         self.cmd('containerapp env java-component spring-cloud-eureka delete -g {} -n {} --environment {} --yes'.format(resource_group, eureka_name, env_name), expect_failure=False)
 
@@ -112,7 +114,6 @@ class ContainerappJavaComponentTests(ScenarioTest):
         java_component_list = self.cmd("containerapp env java-component list -g {} --environment {}".format(resource_group, env_name)).get_output_in_json()
         self.assertTrue(len(java_component_list) == 0)
 
-    @AllowLargeResponse(8192)
     @ResourceGroupPreparer()
     def test_containerapp_java_component(self, resource_group):
         # type "linkers" is not available in North Central US (Stage), if the TEST_LOCATION is "northcentralusstage", use francecentral as location
@@ -125,8 +126,11 @@ class ContainerappJavaComponentTests(ScenarioTest):
         ca_name = self.create_random_name(prefix='javaapp1', length=24)
         config_name = "myconfig"
         eureka_name = "myeureka"
+        sba_name = "mysba"
 
         create_containerapp_env(self, env_name, resource_group)
+        env = self.cmd('containerapp env show -g {} -n {}'.format(resource_group, env_name)).get_output_in_json()
+        default_domain=env['properties']['defaultDomain']
 
         # List Java Components
         java_component_list = self.cmd("containerapp env java-component list -g {} --environment {}".format(resource_group, env_name)).get_output_in_json()
@@ -144,14 +148,27 @@ class ContainerappJavaComponentTests(ScenarioTest):
                 JMESPathCheck('name', eureka_name),
                 JMESPathCheck('properties.componentType', "SpringCloudEureka"),
                 JMESPathCheck('properties.provisioningState', "Succeeded"),
+                JMESPathCheck('properties.ingress.fqdn', eureka_name + "-azure-java.ext." + default_domain),
                 JMESPathCheck('length(properties.configurations)', 2)
             ])
 
         # List Java Components
         java_component_list = self.cmd("containerapp env java-component list -g {} --environment {}".format(resource_group, env_name)).get_output_in_json()
         self.assertTrue(len(java_component_list) == 2)
+     
+        # Create SBA and bind with eureka
+        self.cmd('containerapp env java-component admin-for-spring create -g {} -n {} --environment {} --configuration '.format(resource_group, sba_name, env_name), checks=[
+            JMESPathCheck('name', sba_name),
+            JMESPathCheck('properties.componentType', "SpringBootAdmin"),
+            JMESPathCheck('properties.ingress.fqdn', sba_name + "-azure-java.ext." + default_domain),
+            JMESPathCheck('properties.provisioningState', "Succeeded")
+        ])
+   
+        # List Java Components
+        java_component_list = self.cmd("containerapp env java-component list -g {} --environment {}".format(resource_group, env_name)).get_output_in_json()
+        self.assertTrue(len(java_component_list) == 3)
 
-        # Update Config & Eureka
+        # Update Java Components
         self.cmd(
             'containerapp env java-component config-server-for-spring update -g {} -n {} --environment {} --configuration spring.cloud.config.server.git.uri=https://github.com/Azure-Samples/piggymetrics-config.git'.format(
                 resource_group, config_name, env_name), checks=[
@@ -163,10 +180,18 @@ class ContainerappJavaComponentTests(ScenarioTest):
         self.cmd('containerapp env java-component eureka-server-for-spring update -g {} -n {} --environment {} --configuration'.format(resource_group, eureka_name, env_name), checks=[
                 JMESPathCheck('name', eureka_name),
                 JMESPathCheck('properties.componentType', "SpringCloudEureka"),
+                JMESPathCheck('properties.ingress.fqdn', eureka_name + "-azure-java.ext." + default_domain),
+                JMESPathCheck('properties.provisioningState', "Succeeded")
+        ])
+        self.cmd('containerapp env java-component admin-for-spring update -g {} -n {} --environment {} --bind {}:myeureka --configuration'.format(resource_group, sba_name, env_name, eureka_name), checks=[
+                JMESPathCheck('name', sba_name),
+                JMESPathCheck('properties.componentType', "SpringBootAdmin"),
+                JMESPathCheck('properties.serviceBinds[0].name', eureka_name),
+                JMESPathCheck('properties.ingress.fqdn', sba_name + "-azure-java.ext." + default_domain),
                 JMESPathCheck('properties.provisioningState', "Succeeded")
         ])
 
-        # Show Config & Eureka
+        # Show Java Components
         self.cmd('containerapp env java-component config-server-for-spring show -g {} -n {} --environment {}'.format(resource_group, config_name, env_name), checks=[
             JMESPathCheck('name', config_name),
             JMESPathCheck('properties.componentType', "SpringCloudConfig"),
@@ -176,6 +201,14 @@ class ContainerappJavaComponentTests(ScenarioTest):
         self.cmd('containerapp env java-component eureka-server-for-spring show -g {} -n {} --environment {}'.format(resource_group, eureka_name, env_name), checks=[
             JMESPathCheck('name', eureka_name),
             JMESPathCheck('properties.componentType', "SpringCloudEureka"),
+            JMESPathCheck('properties.ingress.fqdn', eureka_name + "-azure-java.ext." + default_domain),
+            JMESPathCheck('properties.provisioningState', "Succeeded")
+        ])
+        self.cmd('containerapp env java-component admin-for-spring show -g {} -n {} --environment {}'.format(resource_group, sba_name, env_name), checks=[
+            JMESPathCheck('name', sba_name),
+            JMESPathCheck('properties.componentType', "SpringBootAdmin"),
+            JMESPathCheck('properties.ingress.fqdn', sba_name + "-azure-java.ext." + default_domain),
+            JMESPathCheck('properties.serviceBinds[0].name', eureka_name),
             JMESPathCheck('properties.provisioningState', "Succeeded")
         ])
 
@@ -203,9 +236,10 @@ class ContainerappJavaComponentTests(ScenarioTest):
                 JMESPathCheck('length(properties.template.serviceBinds)', 2)
         ])
 
-        # Delete Config & Eureka
+        # Delete Java Components
         self.cmd('containerapp env java-component config-server-for-spring delete -g {} -n {} --environment {} --yes'.format(resource_group, config_name, env_name), expect_failure=False)
-        self.cmd('containerapp env java-component eureka-server-for-spring delete -g {} -n {} --environment {} --yes'.format(resource_group, eureka_name, env_name), expect_failure=False)
+        self.cmd('containerapp env java-component eureka-server-for-spring delete -g {} -n {} --environment {} --yes'.format(resource_group, eureka_name, env_name), expect_failure=False)      
+        self.cmd('containerapp env java-component admin-for-spring delete -g {} -n {} --environment {} --yes'.format(resource_group, sba_name, env_name), expect_failure=False)
 
         # List Java Components
         java_component_list = self.cmd("containerapp env java-component list -g {} --environment {}".format(resource_group, env_name)).get_output_in_json()